[kernel] E: Isolate x86_64 port
Minimize changes to existing x86 source files by extracting all
x86_64-specific code into new dedicated files, selected at compile
time via #[cfg(target_arch)] and #[path] attributes.

New files created:
- arch: idt_x86_64.rs, idtr_x86_64.rs, tss_x86_64.rs, gdtr_x86_64.rs
- nvx: crt0_x86_64.rs (process entry point)
- sys: pm_x86_64.rs (thread start stub)
- uservm: elf64.rs (ELF64 loader), vcpu/reset64.rs (long-mode reset)
- kernel: mm/elf64.rs (kernel ELF64 loader)

Original x86 files (idt.rs, idtr.rs, tss.rs, gdtr.rs) are restored
to their base-commit state with zero modifications.

Also reverts cosmetic changes: dead put_hex() in klog.rs, whitespace
in platform/microvm/mod.rs, and debug logging in vcpu run().

Updates doc/x86_64-port.md with new Code Organization section and
corrected build commands.
