+++ src/libs/arch/src/x86/cpu/idt_x86_64.rs
@@ -16,8 +16,8 @@ pub enum GateType {
     Task32 = 0x5, // 32-bit task gate.
     Int16 = 0x6,  // 16-bit interrupt gate.
     Trap16 = 0x7, // 16-bit trap gate.
-    Int32 = 0xe,  // 32-bit interrupt gate.
-    Trap32 = 0xf, // 32-bit trap gate.
+    Int32 = 0xe,  // 32/64-bit interrupt gate.
+    Trap32 = 0xf, // 32/64-bit trap gate.
 }
 
 //==================================================================================================
@@ -65,34 +65,39 @@ impl From<Flags> for u8 {
 }
 
 //==================================================================================================
-// Interrupt Descriptor Table Entry
+// Interrupt Descriptor Table Entry (64-bit)
 //==================================================================================================
 
-/// Interrupt descriptor table entry (IDTE).
-#[repr(C, align(8))]
+/// Interrupt descriptor table entry (IDTE) for 64-bit x86_64.
+#[repr(C, packed)]
 pub struct Idte {
-    pub handler_low: u16,  // Handler low.
+    pub handler_low: u16,  // Handler bits [0:15].
     pub selector: u16,     // GDT selector.
-    pub zero: u8,          // Always zero.
+    pub ist: u8,           // IST index (bits [0:2]), rest zero.
     pub flags: u8,         // Gate type and flags.
-    pub handler_high: u16, // Handler high.
+    pub handler_mid: u16,  // Handler bits [16:31].
+    pub handler_high: u32, // Handler bits [32:63].
+    pub reserved: u32,     // Must be zero.
 }
 
-// `Idte` must be 8 bytes long. This must match the hardware specification.
-::static_assert::assert_eq_size!(Idte, 8);
+// `Idte` must be 16 bytes long. This must match the hardware specification.
+::static_assert::assert_eq_size!(Idte, 16);
 
 impl Idte {
     /// Creates a new IDT entry.
-    pub fn new(handler: u32, selector: u16, flags: Flags) -> Self {
+    pub fn new(handler: u64, selector: u16, flags: Flags) -> Self {
         let handler_low = handler as u16;
-        let handler_high = (handler >> 16) as u16;
+        let handler_mid = (handler >> 16) as u16;
+        let handler_high = (handler >> 32) as u32;
 
         Self {
             handler_low,
             selector,
-            zero: 0,
+            ist: 0,
             flags: flags.into(),
+            handler_mid,
             handler_high,
+            reserved: 0,
         }
     }
 }
