+++ src/libs/arch/src/x86/mem/gdtr_x86_64.rs
@@ -2,22 +2,22 @@
 // Licensed under the MIT License.
 
 //==================================================================================================
-// Global Descriptor Table Register (GDTR)
+// Global Descriptor Table Register (GDTR) — 64-bit
 //==================================================================================================
 
-/// Global descriptor table register (GDTR).
+/// Global descriptor table register (GDTR) for 64-bit x86_64.
 #[derive(Default)]
 #[repr(C, packed)]
 pub struct Gdtr {
     limit: u16,
-    base: u32,
+    base: u64,
 }
 
-// `Gdtptr` must be 6 bytes long. This must match the hardware specification.
-::static_assert::assert_eq_size!(Gdtr, 6);
+// `Gdtr` must be 10 bytes long. This must match the hardware specification.
+::static_assert::assert_eq_size!(Gdtr, 10);
 
 impl Gdtr {
-    pub fn new(base: u32, size: u16) -> Self {
+    pub fn new(base: u64, size: u16) -> Self {
         Self {
             base,
             limit: size - 1,
